--- Lab1/readme
@@ -1,1 +1,0 @@
-
