+++ Lab2/AppLab2/MainWindow.xaml
@@ -1,0 +1,34 @@
+﻿<Window x:Class="AppLab2.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:AppLab2"
+        mc:Ignorable="d"
+        Title="Genetic Algorithm for Squares" Height="500" Width="800">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <StackPanel Orientation="Horizontal" Grid.Row="0" Margin="5">
+            <Label Content="1x1 Squares:"/>
+            <TextBox Width="50" Margin="5" x:Name="txtSquare1"/>
+            <Label Content="2x2 Squares:"/>
+            <TextBox Width="50" Margin="5" x:Name="txtSquare2"/>
+            <Label Content="3x3 Squares:"/>
+            <TextBox Width="50" Margin="5" x:Name="txtSquare3"/>
+            <Button Content="Start" Margin="5,0" Click="StartButton_Click"/>
+            <Button Content="Stop" Margin="5,0" Click="StopButton_Click"/>
+            <Button Content="Reset" Margin="5,0" Click="ResetButton_Click"/>
+        </StackPanel>
+        <Grid Grid.Row="1">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="1*"/>
+                <ColumnDefinition Width="3*"/>
+            </Grid.ColumnDefinitions>
+            <TextBlock Name="outputText" TextWrapping="Wrap" Margin="5"/>
+            <Canvas Name="canvasRectangles" Grid.Column="1" Background="White" Width="400" Height="400"/>
+        </Grid>
+    </Grid>
+</Window>
