+++ Lab2/GeneticAlg/GeneticAlg.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+	
+    <TargetFramework>net7.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+	
+	<PackageId>EvolutionSquares</PackageId>
+	<Version>1.0.1</Version>
+	<Authors>Guo Haojie</Authors>
+	<Description>A library for square layout using genetic algorithms</Description>
+  </PropertyGroup>
+
+</Project>
